requirements.txt
google-generativeai>=0.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
tabulate>=0.9.0


